converters: all 7 boundary failures fixed -- 5/5 each, and OUT of KNOWN_FAILING
Every one was a MISSING COERCION, not wrong arithmetic:

    kalshi / novig / polymarket   float() before `0.0 < p < 1.0`, which raised
                                  TypeError on '' instead of refusing
    refresh_nba / refresh_wnba    same, and these had NO None guard at all
    ncaaf/prop_model              int("-110.5") RAISES, so a half-point price
                                  refused; float() -> 0.5249406175771971
    report_nfl_props_roi          was 1/5: price=0 returned 0.0, None/'' raised

`report_nfl_props_roi`'s zero is the one worth naming: `abs(0)/(0+100)` is 0.0,
so a MISSING price read as a 0% market rather than as no market -- and 0.0 is a
number a caller will divide an edge by.

REMOVED FROM `KNOWN_FAILING` rather than left sitting in it; that map is now down
to one entry for `probability_to_american`. A tracked-debt list that keeps
entries which have started passing decays into noise the same way the tuple test
did.

THE FALSIFICATION TEST WAS PRE-REGISTERED IN THE LANE AND PASSED: no VALID price
moved. A boundary fix that changes a real conversion is not a boundary fix. All
six probe prices still agree across every implementation (plus_100 0.5 ...
minus_10000 0.9900990099), and the three venue clients still return an identical
-163 for 0.62.

MUTATION-CHECKED one per requirement type, each reverted at RUNTIME so no file
was touched: drop kalshi's string coercion (refuses_empty_string), drop
refresh_wnba's (refuses_none), drop the report_nfl_props_roi zero guard
(refuses_zero), restore ncaaf's int() (accepts_float_price) -- 4 of 4 RED.

CAUGHT BY VERIFYING RATHER THAN TRUSTING A SUCCESS MESSAGE: the first batch
aborted partway and printed the files it HAD patched, so `ncaaf/prop_model` was
silently never edited and sat at 4/5. The re-score found it. A partial success
message is indistinguishable from a complete one unless the outcome is checked.

Verified: 125 passed across the converter files, 115 across the two refresh
runners, 10/10 in the differential.

+++ syndicate/features/ncaaf/prop_model.py
@@ -235,8 +235,12 @@ def american_to_probability(price: Any) -> float | None:
     calling this fair would overstate the market's true probability and make
     every model edge look larger than it is.
     """
+    # FLOAT, NOT INT `[2026-09-05]`. `int("-110.5")` RAISES, so a half-point
+    # price refused instead of converting -- the differential harness probes
+    # -110.5 and expects 0.5249406175. Truncating to -110 would have been worse
+    # than refusing: it returns a plausible number for a price nobody quoted.
     try:
-        value = int(str(price).replace("+", "").strip())
+        value = float(str(price).replace("+", "").strip())
     except (TypeError, ValueError):
         return None
     if value == 0:
